continuity-data-ingestion/AI/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
discord.py==2.4.0
chromadb==0.5.23
sentence-transformers==3.3.1
llama-cpp-python==0.3.16
pydantic==2.10.3
python-dotenv==1.0.1
aiohttp==3.11.11
asyncio-contextmanager==1.0.1
